Add payment redirect support to CheckoutController
`CheckoutController` never had the concept of redirecting the user to the
"offsite" payment method site.
This situation was handled in various bad ways in the past, mostly with the
substitution of the checkout submit button with a link tag, using some JS.

This commit adds an override to CheckoutController to check if the active
payment wants to redirect to its site, before proceeding to the next checkout
step.
Different versions of `#update` overrides are loaded for different Solidus
versions.

The `Payment` class gets a new method called `#redirect_url`.
In this basic implementation it just delegates to the payment method, but this
interface can be used as a base for different setups. For example, some
payment methods may need to use, or even persist, a dynamically generated
redirect URL.