Add PaybrightController
PaybrightController implements the three required endpoints.

`#callback` receives the result of a succesfull Paybright checkout as
parameters, completes the order payment and the order.

`#complete` is where the User is redirected after completing the Paybright
checkout. It also receives the same parameters of `#callback`.

`#cancel` is where the User is redirected after the failure of some Paybright
steps.